Issue 55: Overhaul module structure.
* The rule should not know anything about error messages. This is part of the responsibility of the module calculating rule violations. This module has all information that should be part of the error message and can simply generate it directly without the indirection via Rule.

* Extract all breadth first searches into one file. This makes it easier to find the different search versions.

* Break up large test files.
