+++ tests/integration/interesting_rules_for_tests.py
@@ -1,18 +1,20 @@
-import os
-
 import pytest
-from resources.test_project import src
-from resources.test_project.src import moduleA
-from resources.test_project.src.moduleA import submoduleA1
 
-from pytestarch import EvaluableArchitecture, Rule
-from pytestarch.pytestarch import get_evaluable_architecture
+from pytestarch import Rule
 
-# from matplotlib import pyplot as plt
-# from matplotlib.pyplot import figure
+C = "src.moduleC"
+B = "src.moduleB"
+A = "src.moduleA"
+FILE_A2 = f"{A}.submoduleA2.fileA2"
+A2 = f"{A}.submoduleA2"
+A1 = f"{A}.submoduleA1"
+A11 = f"{A}.submoduleA1.submoduleA11"
+B1_MODULE = f"{B}.submoduleB1"
+B2 = f"{B}.submoduleB1.fileB2"
+FILE_C = f"{C}.fileC"
 
 
-rules_to_test = [
+rules_for_level_limits = [
     (
         Rule()
         .modules_that()
@@ -128,222 +130,7 @@ rules_to_test = [
     ),
 ]
 
-
-@pytest.mark.parametrize("rule, expected_result, skip_with_level_limit", rules_to_test)
-def test_architecture_based_on_string_modules(
-    rule: Rule,
-    expected_result: bool,
-    skip_with_level_limit: bool,
-    graph_based_on_string_module_names: EvaluableArchitecture,
-) -> None:
-    if expected_result:
-        rule.assert_applies(graph_based_on_string_module_names)
-    else:
-        with pytest.raises(AssertionError):
-            assert rule.assert_applies(graph_based_on_string_module_names)
-
-
-@pytest.mark.parametrize("rule, expected_result, skip_with_level_limit", rules_to_test)
-def test_architecture_based_on_module_objects(
-    rule: Rule,
-    expected_result: bool,
-    skip_with_level_limit: bool,
-    graph_based_on_string_module_names: EvaluableArchitecture,
-) -> None:
-    if expected_result:
-        rule.assert_applies(graph_based_on_string_module_names)
-    else:
-        with pytest.raises(AssertionError):
-            assert rule.assert_applies(graph_based_on_string_module_names)
-
-
-def test_depending_on_module_does_not_imply_depending_on_submodule(
-    graph_including_tests: EvaluableArchitecture,
-) -> None:
-    rule_1 = (
-        Rule()
-        .modules_that()
-        .are_named("src.moduleC.cTest")
-        .should()
-        .import_modules_that()
-        .are_named("src.moduleB")
-    )
-    rule_2 = (
-        Rule()
-        .modules_that()
-        .are_named("src.moduleC.cTest")
-        .should_not()
-        .import_modules_that()
-        .are_named("src.moduleB.submoduleB1.fileB3")
-    )
-
-    rule_1.assert_applies(graph_including_tests)
-    rule_2.assert_applies(graph_including_tests)
-
-
-@pytest.mark.parametrize("rule, expected_result, skip_with_level_limit", rules_to_test)
-def test_identical_source_and_module_path_do_not_lead_to_errors(
-    rule: Rule,
-    expected_result: bool,
-    skip_with_level_limit: bool,
-    graph_with_identical_source_and_module_path: EvaluableArchitecture,
-) -> None:
-    if expected_result:
-        rule.assert_applies(graph_with_identical_source_and_module_path)
-    else:
-        with pytest.raises(AssertionError):
-            assert rule.assert_applies(graph_with_identical_source_and_module_path)
-
-
-@pytest.mark.parametrize("rule, expected_result, skip_with_level_limit", rules_to_test)
-def test_level_limit_flattens_dependencies_correctly(
-    rule: Rule,
-    expected_result: bool,
-    skip_with_level_limit: bool,
-    graph_with_level_limit_1: EvaluableArchitecture,
-) -> None:
-    if skip_with_level_limit:
-        return
-
-    if expected_result:
-        rule.assert_applies(graph_with_level_limit_1)
-    else:
-        with pytest.raises(AssertionError):
-            assert rule.assert_applies(graph_with_level_limit_1)
-
-
-# def test_exporting_as_image_does_not_raise_error(
-#     graph_based_on_string_module_names: EvaluableArchitecture,
-# ) -> None:
-#     figure(figsize=(12, 6), dpi=80)
-#
-#     graph_based_on_string_module_names.visualize(
-#         spacing=0.35,
-#         node_size=800,
-#         arrowsize=30,
-#         labels={
-#             "src": "src",
-#             "src.moduleA": "A",
-#             "src.moduleB": "B",
-#             "src.moduleC": "C",
-#             "src.moduleA.submoduleA1": "A1",
-#             "src.moduleA.submoduleA2": "A2",
-#             "src.moduleA.fileA": "fileA",
-#             "src.moduleA.submoduleA1.submoduleA11": "A11",
-#             "src.moduleA.submoduleA1.fileA1": "fileA1",
-#             "src.moduleA.submoduleA1.fileA1_b": "fileA1b",
-#             "src.moduleA.submoduleA1.submoduleA11.fileA11": "fileA11",
-#             "src.moduleA.submoduleA2.fileA2": "fileA2",
-#             "src.moduleB.submoduleB1": "B1",
-#             "src.moduleB.fileB": "fileB",
-#             "src.moduleB.submoduleB1.fileB1": "fileB1",
-#             "src.moduleB.submoduleB1.fileB2": "fileB2",
-#             "src.moduleC.fileC": "fileC",
-#         },
-#     )
-#     plt.savefig("test.png")
-
-
-def test_edges_correctly_calculated_for_level_2_module_path() -> None:
-    level_2_graph = get_evaluable_architecture(
-        os.path.dirname(src.__file__),
-        os.path.dirname(moduleA.__file__),
-        ("*__pycache__", "*__init__.py", "*Test.py"),
-    )
-
-    assert len(level_2_graph._graph._graph.edges) == 0  # no internal dependencies left
-
-
-def test_edges_correctly_calculated_for_level_2_module_path_no_external_dependencies_modified() -> None:
-    level_2_graph = get_evaluable_architecture(
-        os.path.dirname(src.__file__),
-        os.path.dirname(moduleA.__file__),
-        ("*__pycache__", "*__init__.py", "*Test.py"),
-        exclude_external_libraries=False,
-    )
-
-    assert len(level_2_graph._graph._graph.edges) == 16
-
-    for edge in [
-        ("src", "src.moduleA"),
-        ("src", "src.moduleB"),  # must not be src.moduleA.src.moduleB
-        ("src", "src.moduleC"),
-        ("src.moduleA", "src.moduleA.fileA"),
-        ("src.moduleA", "src.moduleA.submoduleA1"),
-        ("src.moduleA", "src.moduleA.submoduleA2"),
-        ("src.moduleA.fileA", "src.moduleC.fileC"),
-        ("src.moduleA.submoduleA1", "src.moduleA.submoduleA1.submoduleA11"),
-        (
-            "src.moduleA.submoduleA1.submoduleA11",
-            "src.moduleA.submoduleA1.submoduleA11.fileA11",
-        ),
-        (
-            "src.moduleA.submoduleA1.submoduleA11.fileA11",
-            "os",
-        ),  # must not be src.moduleA.os!
-        (
-            "src.moduleA.submoduleA1.submoduleA11.fileA11",
-            "src.moduleB.submoduleB1.fileB1",
-        ),
-        ("src.moduleA.submoduleA2", "src.moduleA.submoduleA2.fileA2"),
-        ("src.moduleA.submoduleA2.fileA2", "src.moduleC.fileC"),
-        ("src.moduleB", "src.moduleB.submoduleB1"),
-        ("src.moduleB.submoduleB1", "src.moduleB.submoduleB1.fileB1"),
-        ("src.moduleC", "src.moduleC.fileC"),
-    ]:
-        assert edge in level_2_graph._graph
-
-
-def test_edges_correctly_calculated_for_level_3_module_path() -> None:
-    level_3_graph = get_evaluable_architecture(
-        os.path.dirname(src.__file__),
-        os.path.dirname(submoduleA1.__file__),
-        ("*__pycache__", "*__init__.py", "*Test.py"),
-    )
-
-    assert len(level_3_graph._graph._graph.edges) == 0  # no internal dependencies left
-
-
-def test_edges_correctly_calculated_for_level_3_module_path_no_external_dependencies_modified() -> None:
-    level_3_graph = get_evaluable_architecture(
-        os.path.dirname(src.__file__),
-        os.path.dirname(submoduleA1.__file__),
-        ("*__pycache__", "*__init__.py", "*Test.py"),
-        exclude_external_libraries=False,
-    )
-
-    assert len(level_3_graph._graph._graph.edges) == 7
-
-    for edge in [
-        ("src", "src.moduleB"),
-        ("src.moduleA.submoduleA1", "src.moduleA.submoduleA1.submoduleA11"),
-        (
-            "src.moduleA.submoduleA1.submoduleA11",
-            "src.moduleA.submoduleA1.submoduleA11.fileA11",
-        ),
-        ("src.moduleA.submoduleA1.submoduleA11.fileA11", "os"),
-        (
-            "src.moduleA.submoduleA1.submoduleA11.fileA11",
-            "src.moduleB.submoduleB1.fileB1",
-        ),
-        ("src.moduleB", "src.moduleB.submoduleB1"),
-        ("src.moduleB.submoduleB1", "src.moduleB.submoduleB1.fileB1"),
-    ]:
-        assert edge in level_3_graph._graph
-
-
-C = "src.moduleC"
-B = "src.moduleB"
-A = "src.moduleA"
-FILE_A2 = f"{A}.submoduleA2.fileA2"
-FILE_C = f"{C}.fileC"
-A2 = f"{A}.submoduleA2"
-A1 = f"{A}.submoduleA1"
-A11 = f"{A}.submoduleA1.submoduleA11"
-B1_MODULE = f"{B}.submoduleB1"
-B2 = f"{B}.submoduleB1.fileB2"
-
-test_cases = [
+single_rule_object_error_message_test_cases = [
     pytest.param(
         Rule().modules_that().are_named(C).should().import_modules_that().are_named(A),
         '"src.moduleC" does not import "src.moduleA".',
@@ -921,18 +708,7 @@ test_cases = [
     ),
 ]
 
-
-@pytest.mark.parametrize("rule, expected_error_message", test_cases)
-def test_rule_violated_raises_proper_error_message(
-    rule: Rule,
-    expected_error_message: str,
-    graph_based_on_string_module_names: EvaluableArchitecture,
-) -> None:
-    with pytest.raises(AssertionError, match=expected_error_message):
-        rule.assert_applies(graph_based_on_string_module_names)
-
-
-multiple_rule_objects_test_cases = [
+multiple_rule_objects_error_message_test_cases = [
     pytest.param(
         Rule()
         .modules_that()
@@ -1507,19 +1283,6 @@ multiple_rule_objects_test_cases = [
     ),
 ]
 
-
-@pytest.mark.parametrize(
-    "rule, expected_error_message", multiple_rule_objects_test_cases
-)
-def test_rule_violated_with_multiple_rule_objects_raises_proper_error_message(
-    rule: Rule,
-    expected_error_message: str,
-    graph_based_on_string_module_names: EvaluableArchitecture,
-) -> None:
-    with pytest.raises(AssertionError, match=expected_error_message):
-        rule.assert_applies(graph_based_on_string_module_names)
-
-
 partial_name_match_test_cases = [
     pytest.param(
         Rule()
@@ -1622,17 +1385,3 @@ partial_name_match_test_cases = [
         id="submodule should only be imported by except named - forbidden import - partial name match object",
     ),
 ]
-
-
-@pytest.mark.parametrize("rule, violation", partial_name_match_test_cases)
-def test_rule_violation_correctly_detected_for_partial_name_matches(
-    rule: Rule,
-    violation: bool,
-    graph_based_on_string_module_names: EvaluableArchitecture,
-) -> None:
-    if violation:
-        with pytest.raises(AssertionError):
-            rule.assert_applies(graph_based_on_string_module_names)
-    else:
-        rule.assert_applies(graph_based_on_string_module_names)
-        # no exception
