+++ tests/rule_assessment/test_RuleViolationMessageGenerator.py
@@ -3,9 +3,9 @@ from typing import Dict, List
 import pytest
 
 from pytestarch.eval_structure.evaluable_architecture import (
-    LaxDependenciesByBaseModule,
+    DependenciesByBaseModules,
     Module,
-    StrictDependenciesByBaseModules,
+    UnexpectedDependenciesByBaseModule,
 )
 from pytestarch.rule_assessment.error_message.message_generator import (
     RuleViolationMessageGenerator,
@@ -195,12 +195,12 @@ def test_rule_violation_message_content(
     generator: RuleViolationMessageGenerator,
     violation: Dict[str, bool],
     expected_message: str,
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     violations = RuleViolations(
-        strict_dependencies=strict_dependencies,
-        lax_dependencies=lax_dependencies,
+        dependencies=strict_dependencies,
+        unexpected_dependencies=lax_dependencies,
         **violation,
     )
     messages = generator.create_rule_violation_messages(violations)
@@ -362,12 +362,12 @@ message_only_present_if_rule_violated_test_cases = [
 def test_violated_message_only_present_if_rule_actually_violated(
     generator: RuleViolationMessageGenerator,
     violation: Dict[str, bool],
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     violations = RuleViolations(
-        strict_dependencies=strict_dependencies,
-        lax_dependencies=lax_dependencies,
+        dependencies=strict_dependencies,
+        unexpected_dependencies=lax_dependencies,
         **violation,
     )
     messages = generator.create_rule_violation_messages(violations)
@@ -451,12 +451,12 @@ def test_multiple_rule_objects_combined_in_one_message(
     generator: RuleViolationMessageGenerator,
     violation: Dict[str, bool],
     expected_message: str,
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     violations = RuleViolations(
-        strict_dependencies=strict_dependencies,
-        lax_dependencies=lax_dependencies,
+        dependencies=strict_dependencies,
+        unexpected_dependencies=lax_dependencies,
         **violation,
     )
     messages = generator.create_rule_violation_messages(violations)
@@ -618,12 +618,12 @@ def test_multiple_rule_objects_in_multiple_message(
     violation: Dict[str, bool],
     expected_message_count: int,
     expected_messages: List[str],
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     violations = RuleViolations(
-        strict_dependencies=strict_dependencies,
-        lax_dependencies=lax_dependencies,
+        dependencies=strict_dependencies,
+        unexpected_dependencies=lax_dependencies,
         **violation,
     )
     messages = generator.create_rule_violation_messages(violations)
@@ -669,12 +669,12 @@ forbidden_and_no_import_test_cases = [
 def test_multiple_messages_if_forbidden_and_no_import_both_present(
     violation: Dict[str, bool],
     expected_messages: List[str],
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     violations = RuleViolations(
-        strict_dependencies=strict_dependencies,
-        lax_dependencies=lax_dependencies,
+        dependencies=strict_dependencies,
+        unexpected_dependencies=lax_dependencies,
         **violation,
     )
     messages = single_object_import_generator.create_rule_violation_messages(violations)
