+++ tests/rule_assessment/test_rule_matcher.py
@@ -4,9 +4,9 @@ import pytest
 
 from pytestarch import Rule
 from pytestarch.eval_structure.evaluable_architecture import (
-    LaxDependenciesByBaseModule,
+    DependenciesByBaseModules,
     Module,
-    StrictDependenciesByBaseModules,
+    UnexpectedDependenciesByBaseModule,
 )
 from pytestarch.eval_structure.evaluable_graph import EvaluableArchitectureGraph
 from pytestarch.eval_structure.networkxgraph import NetworkxGraph
@@ -656,8 +656,8 @@ def test_multiple_violations_due_to_multiple_rule_objects(
     imports: List[AbsoluteImport],
     rule: Rule,
     expected_violated: List[str],
-    strict_dependencies: StrictDependenciesByBaseModules,
-    lax_dependencies: LaxDependenciesByBaseModule,
+    strict_dependencies: DependenciesByBaseModules,
+    lax_dependencies: UnexpectedDependenciesByBaseModule,
 ) -> None:
     evaluable = EvaluableArchitectureGraph(NetworkxGraph(ALL_MODULES, imports))
     matcher = rule._prepare_rule_matcher()
@@ -672,5 +672,5 @@ def test_multiple_violations_due_to_multiple_rule_objects(
         else:
             assert not violated
 
-    assert violations.strict_dependencies == strict_dependencies
-    assert violations.lax_dependencies == lax_dependencies
+    assert violations.dependencies == strict_dependencies
+    assert violations.unexpected_dependencies == lax_dependencies
